Rate_map has to take the -start option into account
SINGER takes -start which allows you to run on a segment of the
genome other than starting at basepair=0. Previously the Rate_map
ignored this, meaning it was emitting recombination/mutation rates
for the wrong region. It could also end up producing queries for
coordinates that were prior to the start of the rate map, due to
mutations being offset incorrectly.