Disable "Auto-debug", make parallel_singer robust to failures
Auto-debug doesn't seem to help very often, and is very slow (100
reps is a lot). Disable it.

By default, singer_master does not exit(1) on failure. This is kind
of yucky, but it emits a warning message for now. The easiest way
to get parallel_singer to not terminate after a single segment
fails.